refactor(analyze): extract tier1.ts and tier2.ts from index.ts (#7)
Move heuristic analysis (countMatches, analyzeTier1) to tier1.ts and
AI semantic analysis (ANALYSIS_PROMPT, analyzeTier2) to tier2.ts.
Update index.ts to import from these modules and remove unused imports.
